chore: merge contact utils
