+++ src/frontend/src/lib/utils/contact.utils.ts
@@ -1,4 +1,4 @@
-import type { Contact, ContactImage } from '$declarations/backend/backend.did';
+import type { Contact, ContactImage, ImageMimeType } from '$declarations/backend/backend.did';
 import { updateContact as apiUpdateContact } from '$lib/api/backend.api';
 import { TokenAccountIdSchema } from '$lib/schema/token-account-id.schema';
 import type { Address, OptionAddress } from '$lib/types/address';
@@ -6,7 +6,6 @@ import type { ContactAddressUi, ContactUi } from '$lib/types/contact';
 import type { NetworkId } from '$lib/types/network';
 import type { NonEmptyArray } from '$lib/types/utils';
 import { areAddressesEqual, areAddressesPartiallyEqual } from '$lib/utils/address.utils';
-import { dataUrlToImage } from '$lib/utils/contact-image.utils';
 import {
 	getDiscriminatorForTokenAccountId,
 	getTokenAccountIdAddressString
@@ -133,6 +132,34 @@ export const getNetworkContactKey = ({
 	contact: ContactUi;
 	address: Address;
 }): string => `${address}-${contact.id.toString()}`;
+
+const parseDataUrl = (dataUrl: string): { mime: string; data: Uint8Array } => {
+	const [header, b64] = dataUrl.split(',');
+
+	const match = header.match(/data:(.*);base64/);
+	if (!match) {
+		throw new Error(`Invalid data URL: ${header}`);
+	}
+	const [, mime] = match;
+
+	const bin = atob(b64);
+	const arr = new Uint8Array(bin.length);
+	for (let i = 0; i < bin.length; i++) {
+		arr[i] = bin.charCodeAt(i);
+	}
+	return { mime, data: arr };
+};
+export const dataUrlToImage = (dataUrl: string): ContactImage => {
+	const { mime, data } = parseDataUrl(dataUrl);
+	const mimeType = { [mime]: null } as ImageMimeType;
+	return { mime_type: mimeType, data };
+};
+
+export const imageToDataUrl = (img: ContactImage): string => {
+	const [mime] = Object.keys(img.mime_type);
+	const b64 = btoa(String.fromCharCode(...img.data));
+	return `data:${mime};base64,${b64}`;
+};
 
 export const saveContact = async (params: SaveContactParams): Promise<void> => {
 	const { imageUrl, ...rest } = params;
